# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "BandExplanation": {
        "additionalProperties": false,
        "description": "Answers 'why this band?' for the queue tooltip and the case header.",
        "properties": {
          "band": {
            "$ref": "#/components/schemas/PriorityBand"
          },
          "basis": {
            "description": "How the band was reached, in working language.",
            "title": "Basis",
            "type": "string"
          },
          "caps_applied": {
            "default": [],
            "description": "Guards that limited the band, e.g. text similarity alone cannot reach the top band, or an incomplete bundle lowered certainty.",
            "items": {
              "type": "string"
            },
            "title": "Caps Applied",
            "type": "array"
          }
        },
        "required": [
          "band",
          "basis"
        ],
        "title": "BandExplanation",
        "type": "object"
      },
...
      "CounterEvidenceDto": {
        "additionalProperties": false,
        "description": "One fact that argues against the reason it accompanies, **in words**.\n\nThe refs alone were shipped for a while and it hollowed the feature out: a reviewer saw\n`Encounter ENC-PH-1` under a \"counter-evidence\" heading with nothing saying why it weakened\nanything. The sentence is the argument; the references are only where to look next.",
        "properties": {
          "note": {
            "description": "Working language, written by the rule that raised the reason.",
            "title": "Note",
            "type": "string"
          },
          "refs": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/EvidenceRefDto"
            },
            "title": "Refs",
            "type": "array"
          }
        },
        "required": [
          "note"
        ],
        "title": "CounterEvidenceDto",
        "type": "object"
      },
...
      "ErrorCode": {
        "description": "Stable identifiers. Append new members; never repurpose an existing one.",
        "enum": [
          "BUNDLE_TOO_LARGE",
          "BUNDLE_UNSUPPORTED_CONTENT_TYPE",
          "BUNDLE_DEPTH_EXCEEDED",
          "BUNDLE_MALFORMED_JSON",
          "BUNDLE_SCHEMA_INVALID",
          "BUNDLE_UNKNOWN_RESOURCE_TYPE",
          "BUNDLE_DANGLING_REFERENCE",
          "BUNDLE_CIRCULAR_REFERENCE",
          "BUNDLE_DUPLICATE_RESOURCE_ID",
          "BUNDLE_TOTAL_MISMATCH",
          "INGESTION_NOT_FOUND",
          "CASE_NOT_FOUND",
          "EVALUATION_RUN_NOT_FOUND",
          "BUNDLE_NOT_SCREENABLE",
          "DISPOSITION_REASON_REQUIRED",
          "CASE_VERSION_CONFLICT",
          "DISPOSITION_INVALID_TRANSITION",
          "AUDIT_FORBIDDEN",
          "CASE_ACCESS_FORBIDDEN",
          "CASE_REOPEN_FORBIDDEN",
          "USER_MANAGEMENT_FORBIDDEN",
          "SESSION_INVALID_CREDENTIALS",
          "SESSION_ACCOUNT_DEACTIVATED",
          "USER_NOT_FOUND",
          "USER_SELF_MODIFICATION_REFUSED",
          "USER_NO_CHANGE_REQUESTED",
          "BRIEFING_UNAVAILABLE"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "description": "The single error envelope for every endpoint.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "issues": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            },
            "title": "Issues",
            "type": "array"
          }
        },
        "required": [
          "code",
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "EvidenceRefDto": {
        "additionalProperties": false,
        "description": "A pointer the UI must be able to open.\n\nA reference that does not resolve is a defect, not an empty panel \u2014 see\n`docs/canonical/06_evaluation_plan.md` \u00a7 Evidence-reference validity.",
        "properties": {
          "label": {
            "description": "Short working-language description of the resource.",
            "title": "Label",
            "type": "string"
          },
          "resource_id": {
            "title": "Resource Id",
            "type": "string"
          },
          "resource_type": {
            "$ref": "#/components/schemas/ResourceType"
          }
        },
        "required": [
          "resource_type",
          "resource_id",
          "label"
        ],
        "title": "EvidenceRefDto",
        "type": "object"
      },
...
      "IngestBundleResponse": {
        "additionalProperties": false,
        "description": "Result of submitting one bundle.",
        "properties": {
          "completeness_notes": {
            "default": [],
            "description": "Supporting resources absent from the submission. Travels with the case and lowers certainty downstream rather than raising a signal.",
            "items": {
              "type": "string"
            },
            "title": "Completeness Notes",
            "type": "array"
          },
          "existing_case_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Set when this exact payload and engine version were already screened.",
            "title": "Existing Case Id"
          },
          "ingestion_id": {
            "title": "Ingestion Id",
            "type": "string"
          },
          "input_hash": {
            "description": "SHA-256 of the canonical payload. Drives idempotency.",
            "title": "Input Hash",
            "type": "string"
          },
          "is_screenable": {
            "description": "False when status is INVALID; the UI disables the button.",
            "title": "Is Screenable",
            "type": "boolean"
          },
          "issues": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            },
            "title": "Issues",
            "type": "array"
          },
          "resource_counts": {
            "items": {
              "$ref": "#/components/schemas/ResourceCount"
            },
            "title": "Resource Counts",
            "type": "array"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ValidationStatus"
          }
        },
        "required": [
          "ingestion_id",
          "status",
          "input_hash",
          "resource_counts",
          "is_screenable",
          "schema_version"
        ],
        "title": "IngestBundleResponse",
        "type": "object"
      },
...
      "PriorityBand": {
        "description": "Queue ordering bands.\n\n`NO_OBSERVED_RISK` means no selected detector fired at this engine version. It must never\nbe rendered as \"clean\" or \"safe\" \u2014 the system is not entitled to that claim.",
        "enum": [
          "DETERMINISTIC_CONFLICT",
          "HIGH_PRIORITY_SIGNAL",
          "NEEDS_CONTEXT",
          "NO_OBSERVED_RISK"
        ],
        "title": "PriorityBand",
        "type": "string"
      },
...
      "ReasonCode": {
        "description": "Stable identifiers used by the UI and by tests. Never renumbered.",
        "enum": [
          "LINE_WITHOUT_COMPLETED_PROCEDURE",
          "LINE_WITHOUT_MEDICATION_DISPENSE",
          "SUPPORTING_EVIDENCE_ENTERED_IN_ERROR",
          "OVERLAPPING_CLAIM_SAME_EPISODE",
          "DUPLICATE_CLAIM_FINGERPRINT",
          "NEAR_DUPLICATE_DOCUMENTATION",
          "EPISODE_SPLIT_ACROSS_CLAIMS"
        ],
        "title": "ReasonCode",
        "type": "string"
      },
      "ReasonDto": {
        "additionalProperties": false,
        "description": "One reason, with its evidence and whatever argues against it.\n\n`counter_evidence` is a first-class field rather than a separate lookup: the reviewer must\nsee what weakens a signal on the same screen as the signal itself.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ReasonCode"
          },
          "component_scores": {
            "additionalProperties": {
              "type": "number"
            },
            "description": "Per-component scores, kept alongside the reason rather than collapsed.",
            "title": "Component Scores",
            "type": "object"
          },
          "counter_evidence": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/EvidenceRefDto"
            },
            "title": "Counter Evidence",
            "type": "array"
          },
          "counter_evidence_notes": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/CounterEvidenceDto"
            },
            "title": "Counter Evidence Notes",
            "type": "array"
          },
          "deterministic": {
            "title": "Deterministic",
            "type": "boolean"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceRefDto"
            },
            "title": "Evidence",
            "type": "array"
          },
          "expected_support": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ResourceType"
            },
            "title": "Expected Support",
            "type": "array"
          },
          "mode": {
            "$ref": "#/components/schemas/RiskMode"
          },
          "ruleset_version": {
            "title": "Ruleset Version",
            "type": "string"
          },
          "sentence": {
            "description": "From the reason catalog. Clients never compose their own.",
            "title": "Sentence",
            "type": "string"
          }
        },
        "required": [
          "code",
          "mode",
          "sentence",
          "deterministic",
          "evidence",
          "ruleset_version"
        ],
        "title": "ReasonDto",
        "type": "object"
      },
      "ResourceCount": {
        "additionalProperties": false,
        "properties": {
          "count": {
            "minimum": 0.0,
            "title": "Count",
            "type": "integer"
          },
          "resource_type": {
            "title": "Resource Type",
            "type": "string"
          }
        },
        "required": [
          "resource_type",
          "count"
        ],
        "title": "ResourceCount",
        "type": "object"
      },
      "ResourceType": {
        "enum": [
          "Claim",
          "ClaimLine",
          "Encounter",
          "Condition",
          "Procedure",
          "Medication",
          "Diagnostic",
          "Document",
          "Account",
          "ChargeItem",
          "Invoice",
          "Episode",
          "Practitioner"
        ],
        "title": "ResourceType",
        "type": "string"
      },
      "RiskMode": {
        "description": "The four officially listed facility risk modes this prototype addresses.",
        "enum": [
          "PHANTOM_OR_NO_PROCEDURE_EVIDENCE",
          "REPEAT_BILLING",
          "CLONED_DOCUMENTATION",
          "UNBUNDLING_FRAGMENTATION"
        ],
        "title": "RiskMode",
        "type": "string"
      },
...
      "ScreenRequest": {
        "additionalProperties": false,
        "description": "Screening carries no options.\n\nThere is deliberately no detector, threshold, or mode selection \u2014 the ingest screen offers\none button, and a configuration wizard would let a presenter tune their way to a result.",
        "properties": {
          "engine_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Pin a version for reproducibility. Defaults to current.",
            "title": "Engine Version"
          }
        },
        "title": "ScreenRequest",
        "type": "object"
      },
      "ScreenResponse": {
        "additionalProperties": false,
        "description": "Result of screening one ingested bundle.",
        "properties": {
          "band": {
            "$ref": "#/components/schemas/BandExplanation"
          },
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "case_version": {
            "description": "Required by the disposition endpoint.",
            "minimum": 1.0,
            "title": "Case Version",
            "type": "integer"
          },
          "latency_ms": {
            "minimum": 0.0,
            "title": "Latency Ms",
            "type": "integer"
          },
          "primary_reason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReasonDto"
              },
              {
                "type": "null"
              }
            ],
            "description": "Strongest reason, opened first by the case detail screen."
          },
          "reasons": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ReasonDto"
            },
            "title": "Reasons",
            "type": "array"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "versions": {
            "$ref": "#/components/schemas/VersionStamp"
          }
        },
        "required": [
          "case_id",
          "case_version",
          "state",
          "band",
          "versions",
          "latency_ms"
        ],
        "title": "ScreenResponse",
        "type": "object"
      },
...
      "ValidationIssue": {
        "additionalProperties": false,
        "description": "One problem found in a submitted bundle, pointed at a specific resource.\n\n`resource_type` and `resource_id` are what make the ingest screen's error list\nactionable \u2014 the operator learns *which* resource to fix, not merely that something is\nwrong somewhere.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "detail": {
            "description": "Human-readable explanation. Never contains medical text.",
            "title": "Detail",
            "type": "string"
          },
          "resource_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Id"
          },
          "resource_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Type"
          }
        },
        "required": [
          "code",
          "detail"
        ],
        "title": "ValidationIssue",
        "type": "object"
      },
      "ValidationStatus": {
        "description": "Three outcomes, not two.\n\n`VALID_WITH_NOTES` exists because an incomplete record and a billed-but-unevidenced\nservice look identical at the schema level. Collapsing them into \"valid\" or \"invalid\" is\nhow this system would produce false accusations, so the distinction is in the type.",
        "enum": [
          "VALID",
          "VALID_WITH_NOTES",
          "INVALID"
        ],
        "title": "ValidationStatus",
        "type": "string"
      },
      "VersionStamp": {
        "additionalProperties": false,
        "description": "Engine identity carried on every screening result and audit event.",
        "properties": {
          "dataset_version": {
            "title": "Dataset Version",
            "type": "string"
          },
          "engine_version": {
            "title": "Engine Version",
            "type": "string"
          },
          "ruleset_version": {
            "title": "Ruleset Version",
            "type": "string"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "ruleset_version",
          "engine_version",
          "dataset_version"
        ],
        "title": "VersionStamp",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Claim evidence integrity screening. Synthetic data only.",
    "title": "TilikKlaim API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/bundles": {
      "post": {
        "description": "Validate a bundle and return counts, issues, and a deterministic input hash.",
        "operationId": "ingest_bundle_v1_bundles_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Actor-Role",
            "required": false,
            "schema": {
              "default": "reviewer",
              "title": "X-Actor-Role",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestBundleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Submit one synthetic bundle for validation",
        "tags": [
          "bundles"
        ]
      }
    },
    "/v1/bundles/{ingestion_id}/screen": {
      "post": {
        "description": "Screen a validated bundle and return its reasons, band, and case.\n\nIdempotent for the same input hash and engine version: re-screening returns the existing\ncase with its version bumped, rather than creating a second case for one claim.",
        "operationId": "screen_ingested_bundle_v1_bundles__ingestion_id__screen_post",
        "parameters": [
          {
            "in": "path",
            "name": "ingestion_id",
            "required": true,
            "schema": {
              "title": "Ingestion Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-Role",
            "required": false,
            "schema": {
              "default": "reviewer",
              "title": "X-Actor-Role",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScreenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScreenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Screen an ingested bundle",
        "tags": [
          "bundles"
        ]
      }
    },
...
  }
}